requirements.txt
africastalking==1.2.3
amqp==2.6.1
attrs==20.3.0
autopep8==1.5.4
billiard==3.6.3.0
boto3==1.12.46
botocore==1.15.49
celery==4.4.2
certifi==2020.12.5
chardet==3.0.4
contextlib2==0.6.0.post1
coverage==5.3
Django==3.2
django-celery-beat==1.6.0
django-cors-headers==3.6.0
django-filter==2.2.0
django-redis==4.12.1
django-timezone-field==3.1
djangorestframework==3.12.4
docopt==0.6.2
docutils==0.15.2
Faker==0.9.1
fcm-django==0.3.7
flower==0.9.4
gevent==20.9.0
greenlet==0.4.17
gunicorn==20.0.4
humanize==0.5.1
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
jmespath==0.10.0
kombu==4.6.11
mixer==6.1.3
num2words==0.5.10
packaging==20.8
paystackapi==2.0.0
phonenumbers==8.11.4
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.10.0
pycodestyle==2.6.0
pyfcm==1.4.9
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.0
pytest-cov==2.10.1
pytest-django==3.9.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-http-client==3.3.1
pytz==2020.4
redis==3.5.3
requests==2.25.0
s3transfer==0.3.3
schema==0.7.2
sendgrid==6.4.8
sentry-sdk==0.14.3
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
text-unidecode==1.2
toml==0.10.2
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.25.11
vine==1.3.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
